fix: Move data-converter out of runtime path
Fixes #1.
