+++ src/data-mbta-parking.json.js
@@ -1,6 +1,6 @@
 const config = require('./config.js');
 const fetch = require('node-fetch');
-const {getGeoJson} = require('./converter.js');
+const {getGeoJson} = require('./data-converter.js');
 
 const JSON_URL = "https://api-v3.mbta.com/facilities?filter%5Btype%5D=PARKING_AREA&page%5Blimit%5D=5000&include=stop";
