+++ src/index.js
@@ -1,7 +1,7 @@
 import "./style.css";
 import bubbleImage from './bubble.png';
 import config from './config.js';
-import {getColorInterpolateArray, QUANTILE_STOPS_GTR, QUANTILE_STOPS_RTG} from './converter';
+import {getColorInterpolateArray, QUANTILE_STOPS_GTR, QUANTILE_STOPS_RTG} from './colorGen';
 
 const mbtaParkingJsonDownload = require('./data-mbta-parking.json.js');
 const mbtaLinesJsonDownload = require('./data-mbta-lines.json.js');
